Schedule[Controller, service, domain] 이름변경 => Board, 계획 일정 조회 Plan java단으로 이동
