+++ JejuPlanner/src/main/java/com/gteam/planner/dao/BoardDAO.java
@@ -1,0 +1,9 @@
+package com.gteam.planner.dao;
+
+
+
+public interface BoardDAO {
+	
+
+
+}
