+++ JejuPlanner/src/main/java/com/gteam/planner/dao/BoardDAOImpl.java
@@ -1,0 +1,18 @@
+package com.gteam.planner.dao;
+
+
+import org.apache.ibatis.session.SqlSession;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Repository;
+
+
+@Repository
+public class BoardDAOImpl implements BoardDAO{
+	
+	@Autowired
+	private SqlSession sql;
+	private static String namespace = "com.gteam.planner.mappers.schedule";
+	
+	
+	
+}
