+++ JejuPlanner/src/main/java/com/gteam/planner/service/BoardService.java
@@ -1,0 +1,10 @@
+package com.gteam.planner.service;
+
+
+public interface BoardService {
+	
+	
+	// 일정 객체
+	/*public ScheduleVO formData() throws Exception;*/
+
+}
