+++ JejuPlanner/src/main/java/com/gteam/planner/service/BoardServiceImpl.java
@@ -1,0 +1,19 @@
+package com.gteam.planner.service;
+
+
+import org.springframework.stereotype.Service;
+
+
+@Service
+public class BoardServiceImpl implements BoardService{
+	
+	
+
+	
+	//일정 객체 출력
+/*	@Override
+	public ScheduleVO formData() throws Exception {
+		return dao.list().get(0);
+	}*/
+
+}
