+++ JejuPlanner/src/main/java/com/gteam/planner/service/PlanService.java
@@ -17,8 +17,12 @@ public interface PlanService {
 	
 	//계획 설정 추가, PlanNo 확인, 계획 DB 추가
 	public void planAdd(PlanVO vo, List<Map<String,Object>> schList) throws Exception;
+	
 	//게시판용 계획 목록
 	List<PlanVO> planList() throws Exception;
+	
+	//계획일정 목록
+	public List<ScheduleVO> planSchList(int planNo) throws Exception;
 
 	//유저별 계획 목록
 	List<PlanVO> planListForUser(String userId) throws Exception;
