--- JejuPlanner/src/main/java/com/gteam/planner/service/ScheduleService.java
@@ -1,15 +1,0 @@
-package com.gteam.planner.service;
-
-import java.util.List;
-
-import com.gteam.planner.domain.ScheduleVO;
-
-public interface ScheduleService {
-	
-	// 일정 목록
-	public List<ScheduleVO> scheduleList(int planNo) throws Exception;
-	
-	// 일정 객체
-	/*public ScheduleVO formData() throws Exception;*/
-
-}
