--- JejuPlanner/src/main/resources/mappers/scheduleMapper.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper
-  PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-  "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.gteam.planner.mappers.schedule">
-
-		<!-- 일정 조회 -->
-		<select id="scheduleList" parameterType="int" resultType="com.gteam.planner.domain.ScheduleVO">
-			SELECT planNo, userId, descript, addr, startTime, planDay
-			FROM schedule
-			WHERE planNo = #{planNo}
-		</select>
-
-</mapper>
